relocated user login and sign up to sub screens
