--- src/screens/App/screens/Login/index.js
@@ -1,3 +1,0 @@
-import Login from "./components/Login";
-
-export default Login;
