+++ src/screens/App/screens/Login/route.js
@@ -1,14 +1,11 @@
-import routes from "../../../../shared/Route/routes";
-import Login from "./index";
-
 // subroutes
 import SellerLoginRoute from "./screens/Seller/route";
 import UserLoginRoute from "./screens/User/route";
 
 const route = {
-  path: routes.LOGIN.path,
-  component: Login,
-  exact: true,
+  path: undefined,
+  component: undefined,
+  exact: undefined,
 
   subroutes: {
     SELLER: SellerLoginRoute,
