--- src/screens/App/screens/SignUp/index.js
@@ -1,3 +1,0 @@
-import SignUp from "./components/SignUp";
-
-export default SignUp;
