+++ src/screens/App/screens/SignUp/route.js
@@ -1,14 +1,11 @@
-import routes from "../../../../shared/Route/routes";
-import SignUp from "./index";
-
 // subroutes
 import SellerSignUp from "./screens/Seller/route";
 import UserSignUp from "./screens/User/route";
 
 const route = {
-  path: routes.SIGN_UP.path,
-  component: SignUp,
-  exact: true,
+  path: undefined,
+  component: undefined,
+  exact: undefined,
 
   subroutes: {
     SELLER: SellerSignUp,
